requirements.txt
annotated-types==0.7.0
anyio==3.7.1
bcrypt==4.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.1.0
fastapi==0.104.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
limits==4.4.1
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.6.1
pydantic==2.4.2
pydantic_core==2.10.1
python-jose==3.3.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
six==1.17.0
slowapi==0.1.8
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.23.2
wrapt==1.17.2
